requirements.txt
asn1crypto==0.24.0
blivet==3.1.0
Brlapi==0.6.7
cffi==1.11.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
configobj==5.0.6
configparser==5.0.0
coverage==4.5.1
cryptography==2.3
cupshelpers==1.0
dbus-python==1.2.4
decorator==4.2.1
Django==2.0.1
django-haystack==2.8.1
django-tinymce==3.0.2
gpg==1.10.0
idna==2.5
initial-setup==0.3.62.1
isc==2.0
jieba==0.42.1
langtable==0.0.38
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
netifaces==0.10.6
ntplib==0.3.3
ordered-set==2.0.2
pciutils==2.3.6
perf==0.1
pid==2.1.1
Pillow==7.1.2
ply==3.9
productmd==1.11
pwquality==1.4.0
pycairo==1.16.3
pycparser==2.14
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
pygobject==3.28.3
pykickstart==3.16.7
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pyparted==3.11.0
PySocks==1.6.8
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-linux-procfs==0.6
python-meh==0.47.2
pytz==2017.2
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
requests==2.20.0
requests-file==1.4.3
requests-ftp==0.3.1
rhnlib==2.8.6
rpm==4.14.2
schedutils==0.6
selinux==2.9
sepolicy==1.1
setools==4.2.2
setroubleshoot==1.1
simpleline==1.1
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.6
SSSDConfig==2.2.0
syspurpose==1.25.17
systemd-python==234
urllib3==1.24.2
Whoosh==2.7.4
